move util into providers
